Reorganizing code to MVC
