+++ express-project/controllers/messages.controller.js
@@ -1,0 +1,12 @@
+function getMessages(req, res) {
+    res.send("<ul><li>helloooo</li></ul>");
+};
+
+function postMessages(req, res) {
+    res.send("Updating messages...");
+};
+
+module.exports = {
+    getMessages,
+    postMessages
+};
